services/push-service/requirements.txt
fastapi[standard]
uvicorn
aio-pika
redis[asyncio]
httpx
pydantic
python-dotenv
circuitbreaker
python-multipart
google-auth
requests
pytest
pytest-asyncio

pytest-cov
pytest-mock
faker